requirements.txt
aiofiles
aiohttp
discord.py